+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 2.1.0 Unreleased
 
 - Add the `constant` combinator, typically used to tag objects.
+- Add the `null` combinator.
 
 # 2.0.0 (2024-01-12)