+++ CHANGES.md
@@ -6,6 +6,7 @@
   combinators.
 - Add the `union` combinator.
 - Add the `dft` field combinator.
+- Add the `json` combinator.
 
 # 2.0.0 (2024-01-12)